[BUGFIX] Properly resolve template object fileref
At this point in time, the fileref field of the table
tx_templavoila_tmplobj is not yet fully FAL-compatible.
This means, that the field still contains relative paths
and not identifiers like "file:1234".

However, it is possible to reference a file by a FAL
identifier using the file wizard.

Therefore, the fileref needs to be dynamically resolved.
If a FAL identifier is given, a File object will be
fetched via the FileRepository, otherwise, the path is
considered to be relative. Either actually relative to
the docroot or via the syntax "Ext:foo/path/to/file.html"