[BUGFIX] Do not override new page wizard
Unfortunately the algorithm for resolving the new page
wizard (core) is not very dynamic. This leads to errors
trying to override the new page wizard unless it follows
a decent directory structure.

Therefore, the new page wizard of templavoila will be
disabled for now.

See: https://forge.typo3.org/issues/79370