[BUGFIX] Fix detection of language key in FrontendController
Before TYPO3 7.6, the core relied on static_info_tables to
evaluate sys_language_isocode. If it was installed and the
sys_language_uid was > 0, there was a database lookup for
the current language. The value was uppercase by default
and there was no value for the default language 0 (zero).

This changed with TYPO3 7.6, when the isocode value became
part of the sys_language table. From now on the default
value is "en" and the values in the database are lowecase.

This resulted in two different bugs:

1) The default sheet could not be found any more, because
   only "lDEF" is defined in the flexforms for the default
   language.

2) The language key for languages > 0 were lowercase, thus
   resulting in a mismatch during a string comparison.
   "lEN" !== "len"

Solution to the problem is:
The isocode for the default language "de", will be ignored
and replaced with "DEF". For all other defined languages
the isocode, stored in the database, will be uppercased.