[TASK] Simplify and harden templateObjectItemsProcFunc again
When looking at the code again, I found that calling
BackendUtility::getTCEFORM_TSconfig is not necessary
at all. Thus it is completely removed, along with
the code that prepared the current row for the use
within that method. This does simplify and harden
the whole method and therefore avoids PHP warnings.