[BUGFIX] Ignore storage folder if undefined
Since 7.6.3 there is the possibility to define
a storage folder for the templates and data
structures by PageTS. This is a replacement for
the definition via page properties, that has been
removed in TYPO3 7.6.

Unfortunately that broke the template selection
if there wasn't a storage folder defined at all.

This commit fixes this behaviour. One can either
set a storage folder for filtering the available
templates and data structures or leave it as is
which then makes all records available.